Added ProtectYourEars.h and a multitap delay in AudioProcessor
Added the ProtectYourEars header file to help avoid any catastrophic hearing damage caused by faulty code. As well as adding a multitap delay to add 2 repetitions of delay instead of just one to get a glimpse of what we'll be able to do with the feedback function which will be added next.